SystemCode/Rental-Recommendation-System-in-Singapore/requirements.txt
asent=0.8.0=pypi_0
deap=1.4.1=py311h320fe9a_1
dill=0.3.7=pyhd8ed1ab_0
flask=2.3.3=pyhd8ed1ab_0
flask-bcrypt=1.0.1=pyhd8ed1ab_0
flask-login=0.6.2=pyhd8ed1ab_1
flask-migrate=4.0.5=pyhd8ed1ab_0
flask-minify=0.42=pyhd8ed1ab_0
flask-sqlalchemy=3.0.3=pyhd8ed1ab_0
flask-wtf=1.2.1=pyhd8ed1ab_0
folium=0.14.0=pyhd8ed1ab_0
langdetect=1.0.9=pyhd8ed1ab_0
matplotlib=3.8.0=py311h38be061_2
matplotlib-base=3.8.0=py311h54ef318_2
matplotlib-inline=0.1.6=pyhd8ed1ab_0
numpy=1.26.0=py311h64a7726_0
pandas=2.1.1=py311h320fe9a_1
python-dotenv=1.0.0=pyhd8ed1ab_1
python-flair=0.12.2=pyhd8ed1ab_0
scikit-learn=1.3.1=py311hc009520_1
scikit-surprise=1.1.3=py311h1f0f07a_2
textblob=0.15.3=py_0
wtforms=3.0.1=pyhd8ed1ab_0
